Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main() {
-    double a,b;
-    scanf("%d",&a);
-    scanf("%d",&b);
-    printf("product: %d", a*b);
+    float a,b;
+    scanf("%f",&a);
+    scanf("%f",&b);
+    printf("product: %.2f", a*b);
 }
